Retain typeahead value when clicking away from the field (hacky...)
